% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quant.R
\name{compute_cat}
\alias{compute_cat}
\title{Count or Proportion}
\usage{
compute_cat(x, prop = F, na.rm = T, digits = 3)
}
\arguments{
\item{x}{a vector containing entries for the indicator variable (either TRUE/FALSE or 1/0).}

\item{prop}{a logical. If TRUE, then a proportion is returned. If FALSE, then a count
is returned.}

\item{na.rm}{a logical. If TRUE, then NA values are removed.}

\item{digits}{an integer specifying the number of significant digits to report.}
}
\value{
a scalar containing the count or proportion.
}
\description{
\code{compute_cat} takes an indicator variable and returns either the
proportion or count of TRUE values.
}
