% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quant.R
\name{compute_meansd}
\alias{compute_meansd}
\title{Basic Table, Mean (SD) format.}
\usage{
compute_meansd(x, combine = T, digits = 2)
}
\arguments{
\item{x}{a numeric vector.}

\item{combine}{a logical value. If \code{TRUE}, then the mean and SD are
combined as 'Mean (SD)'. If \code{FALSE}, then the mean and SD are reported
separately.}

\item{digits}{an integer to pass to \code{round}.}

\item{...}{additional arguments to pass to \code{mean} and \code{sd}.}
}
\value{
a character vector containing the computed statistics.
}
\description{
\code{compute_meansd} computes N, Mean and SD.
}
