% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quant.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute Means, SDs, and Sample Sizes}
\usage{
compute_stats(x, labels = NULL)
}
\arguments{
\item{x}{a numeric vector.}

\item{labels}{a vector of character strings containing the labels for
the statistics computed on \code{x}.}
}
\value{
a numeric vector containing the computed statistics.
}
\description{
\code{compute_stats} is a function for computing the mean, SD, and N
for a numeric variable. Designed to be passed as the \code{quant_func}
argument to \code{survey_table}.
}
