% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quant.R
\name{compute_stats_meansdcombined}
\alias{compute_stats_meansdcombined}
\title{Basic Table, Mean (SD) format.}
\usage{
compute_stats_meansdcombined(x, labels = NULL, digits = 2)
}
\arguments{
\item{x}{a numeric vector.}

\item{labels}{a vector of character strings.}

\item{digits}{an integer to pass to \code{round}.}
}
\value{
a character vector containing the computed statistics.
}
\description{
\code{compute_stats_meansdcombined} computes the basic summary statistics
(N, mean, and SD) and combines mean and SD in the format 'Mean (SD)'.
}
