% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quant.R
\name{construct_indicators}
\alias{construct_indicators}
\title{Expand Factor Variable into Indicators}
\usage{
construct_indicators(var_name, dataset)
}
\arguments{
\item{var_name}{a character string or vector containing the name(s) of the factor variable(s).}

\item{dataset}{a data.frame object containing \code{var_name}.}
}
\value{
a data.frame containing the expanded factor variable.
}
\description{
\code{construct_indicators} takes a factor variable with k levels contained in a data.frame and
expands it into a k-column data.frame of indicator variables for each of the levels.
Primarily a helper function for \code{magic_table_categorical}.
}
