% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.magic_table.R
\name{flextable.magic_table}
\alias{flextable.magic_table}
\title{Magic Table Generic for Flextable}
\usage{
flextable.magic_table(data)
}
\arguments{
\item{data}{an object of class \code{magic_table}.}

\item{...}{additional arguments to be passed to the function \code{flextable}.}
}
\description{
Magic Table Generic for Flextable
}
