% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_table.R
\name{is.magic_table}
\alias{is.magic_table}
\title{\code{magic_table} Object?}
\usage{
is.magic_table(x)
}
\arguments{
\item{x}{an \code{R} object.}
}
\value{
a logical indicting whether \code{x} is of class \code{magic_table}.
}
\description{
\code{is.magic_table} checks if the input is of class \code{magic_table}.
}
