% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_table.R
\name{magic_table}
\alias{magic_table}
\title{Magic Table}
\usage{
magic_table(dataset, vars, var_labels, cat_labels, stat_func, condvar_wide,
  condvar_wide_labels, condvar_long, condvar_long_labels, categorical = NULL,
  ...)
}
\arguments{
\item{dataset}{a data.frame object, or object that can be coerced to a data.frame.}

\item{vars}{a vector of character strings containing the names of the variables
for which summary statistics should be computed.}

\item{var_labels}{a vector of character strings containing the labels for the variables
in \code{vars}.}

\item{cat_labels}{a vector of character strings containing the labels for the categories
of the variables contained in \code{vars}. If NULL (default), then \code{cat_labels} is
constructed from the \code{levels} of \code{vars}.}

\item{stat_func}{a function for calculating the summary statistics for the variables specified in \code{vars}.}

\item{condvar_wide}{a character string indicating a variable in \code{dataset} for wide conditioning,
i.e. within-group statistics are added as extra columns.}

\item{condvar_wide_labels}{a character vector containing the labels for the different values
of \code{condvar_wide}.}

\item{condvar_long}{a character string indicating a variable in \code{dataset} for long conditioning,
i.e. within-group statistics are added as extra rows.}

\item{condvar_long_labels}{a character vector containing the labels for the different values
of \code{condvar_long}.}

\item{categorical}{a logical value indicating whether \code{magic_table_categorical} or
\code{magic_table_numeric} should be called. If \code{NULL} (default), then \code{magic_table}
wil automatically choose one based on the classes of \code{vars}.}

\item{...}{additional variables to be passed to \code{stat_func}.}

\item{stat_labels}{a vector of character strings containing the labels for the values
in the output from \code{stat_func}.}
}
\value{
a \code{magic_table} object, the result of a call to either \code{magic_table_numeric}
or \code{magic_table_categorical}.
}
\description{
\code{magic_table} is the top-level function for \code{magic_table_numeric} and
\code{magic_table_categorical}.
}
\details{
The main purpose of the \code{magic_table} function is to serve as a top-level function
for calling either \code{magic_table_numeric} or \code{magic_table_categorical}. For convenience,
\code{magic_table} will attempt to figure out which one should be called based on the classes
of the variables contained in \code{vars}.
}
