% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_table.R
\name{magic_table_categorical}
\alias{magic_table_categorical}
\title{Magic Table}
\usage{
magic_table_categorical(dataset, vars, var_labels = NULL, cat_labels = NULL,
  stat_func, stat_labels, condvar_wide = NULL, condvar_wide_labels = NULL,
  condvar_long = NULL, condvar_long_labels = NULL, ...)
}
\arguments{
\item{dataset}{a data.frame object, or object that can be coerced to a data.frame.}

\item{vars}{a vector of character strings containing the names of the variables
for which summary statistics should be computed.}

\item{var_labels}{a vector of character strings containing the labels for the variables
in \code{vars}.}

\item{cat_labels}{a vector of character strings containing the labels for the categories
of the variables contained in \code{vars}. If NULL (default), then \code{cat_labels} is
constructed from the \code{levels} of \code{vars}.}

\item{stat_func}{a function for calculating the summary statistics for the variables specified in \code{vars}.}

\item{condvar_wide}{a character string indicating a variable in \code{dataset} for wide conditioning,
i.e. within-group statistics are added as extra columns.}

\item{condvar_wide_labels}{a character vector containing the labels for the different values
of \code{condvar_wide}.}

\item{condvar_long}{a character string indicating a variable in \code{dataset} for long conditioning,
i.e. within-group statistics are added as extra rows.}

\item{condvar_long_labels}{a character vector containing the labels for the different values
of \code{condvar_long}.}
}
\value{
a list containing tables of summary statistics for the specified variables.
}
\description{
\code{magic_table_categorical} creates a magic table object from a data.frame-type object when
the input variables are a set of categorical variables.
}
