% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_table.R
\name{magic_table_numeric}
\alias{magic_table_numeric}
\title{Magic Table}
\usage{
magic_table_numeric(dataset, vars, var_labels = NULL, stat_func, stat_labels,
  condvar_wide = NULL, condvar_wide_labels = NULL, condvar_long = NULL,
  condvar_long_labels = NULL, ...)
}
\arguments{
\item{dataset}{a data.frame object, or object that can be coerced to a data.frame.}

\item{vars}{a vector of character strings containing the names of the variables
for which summary statistics should be computed.}

\item{var_labels}{a vector of character strings containing the labels for the variables
contained in \code{vars}.}

\item{stat_func}{a function for calculating the summary statistics for the variables specified in \code{vars}.}

\item{condvar_wide}{a character string indicating a variable in \code{dataset} for wide conditioning,
i.e. within-group statistics are added as extra columns.}

\item{condvar_wide_labels}{a character vector containing the labels for the different values
of \code{condvar_wide}.}

\item{condvar_long}{a character string indicating a variable in \code{dataset} for long conditioning,
i.e. within-group statistics are added as extra rows.}

\item{condvar_long_labels}{a character vector containing the labels for the different values
of \code{condvar_long}.}

\item{...}{additional variables to be passed to \code{stat_func}.}

\item{categorical}{a logical value indicating whether \code{magic_table_numeric} or
\code{magic_table_categorical} should be called. If \code{NULL} (default), then
\code{magic_table} will select one based on the class of the variables contained in \code{vars}.}
}
\value{
a list containing tables of summary statistics for the specified variables.
}
\description{
\code{magic_table_numeric} creates a magic table object from a data.frame-type object when
the input variables are a set of numeric variables.
}
