% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_table.R
\name{survey_table}
\alias{survey_table}
\title{Survey Tables}
\usage{
survey_table(dataset, vars, stat_func, condvar_wide = NULL,
  condvar_long = NULL, ...)
}
\arguments{
\item{dataset}{a data.frame object.}

\item{vars}{a vector of character strings containing the names of the variables
for which summary statistics should be computed.}

\item{stat_func}{a function for calculating the summary statistics for the variables specified in \code{vars}.}

\item{condvar_wide}{a character string indicating a variable in \code{dataset} for wide conditioning,
i.e. within-group statistics are added as extra columns.}

\item{condvar_long}{a character string indicating a variable in \code{dataset} for long conditioning,
i.e. within-group statistics are added as extra rows.}

\item{...}{additional variables to be passed to \code{stat_func}.}
}
\value{
a list containing tables of summary statistics for the specified variables.
}
\description{
\code{survey_table} takes a dataset containing responses to WH
survey questions at different timepoints and constructs raw tables
of summary statistics.
}
