% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_table.R
\name{magic_table}
\alias{magic_table}
\title{Magic Table}
\usage{
magic_table(dataset, vars, var_labels = NULL, stat_func, stat_labels,
  condvar_wide = NULL, condvar_wide_labels = NULL,
  condvar_long = NULL, condvar_long_labels = NULL, ...)
}
\arguments{
\item{dataset}{a data.frame object, or object that can be coerced to a data.frame.}

\item{vars}{a vector of character strings containing the names of the variables
for which summary statistics should be computed.}

\item{stat_func}{a function for calculating the summary statistics for the variables specified in \code{vars}.}

\item{condvar_wide}{a character string indicating a variable in \code{dataset} for wide conditioning,
i.e. within-group statistics are added as extra columns.}

\item{condvar_long}{a character string indicating a variable in \code{dataset} for long conditioning,
i.e. within-group statistics are added as extra rows.}

\item{...}{additional variables to be passed to \code{stat_func}.}
}
\value{
a list containing tables of summary statistics for the specified variables.
}
\description{
\code{magic_table} creates a magic table object from a data.frame-type object.
}
