% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Word-Style Tables}
\usage{
make_table(tables, var_labels, stat_labels, condvar_wide_labels = NULL,
  condvar_long_labels = NULL)
}
\arguments{
\item{tables}{a list of tables of summary statistics. Should be the
output from a call to \code{survey_table}.}

\item{var_labels}{a vector of character strings containing the labels for the variables.}

\item{stat_labels}{a vector of character strings containing the labels for the
summary statistics, e.g. "Mean", "SD".}

\item{condvar_wide_labels}{a vector of character strings containing the
labels for the "wide" conditioning variable, e.g. "Baseline", "6 Months", "6 Month-Baseline Difference".}

\item{condvar_long_labels}{a vector of character strings containing the labels
for the "long" conditioning variable, e.g. "No Treatment", "Treatment".}
}
\value{
a flextable object.
}
\description{
\code{make_table} takes the output from a call to \code{survey_table} and
formats it using the \code{flextable} package.
}
